--- code/05-templating/templates/base.html
@@ -1,13 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-  <head>
-    <title>
-    {% block title %}
-      Index
-    {% endblock %}
-    </title>
-  </head>
-  <body>
-    <div id="content"> {% block content %} {% endblock %} </div>
-  </body>
-</html>
