--- code/05-templating/templates/edit.html
@@ -1,25 +1,0 @@
-{% extends "base.html" %}
-{% block title %}
-Edit Post {{ post.title }}
-{% endblock %}
-
-{% block content %}
-<h1>Edit: {{ post.title }}</h1>
-<form action="/{{ post.id }}/edit" method="POST">
-  <p>
-    Title:
-    <input type="text" name="title" value="{{ post.title }}">
-  </p>
-  <p>Text:
-    <textarea name="text" rows="10">{{ post.text }}</textarea>
-  </p>
-  <input type="submit">
-</form>
-<hr>
-<p><small> created by: {{ owner }}, last editor: {{ editor }} </small></p>
-<p>
-  <a href="/">list</a>
-  <a href="/{{ post.id }}">view</a>
-  <a href="/{{ post.id }}/delete">delete</a>
-</p>
-{% endblock %}
