--- code/05-templating/templates/example.html
@@ -1,15 +1,0 @@
-<html>
-    <head>
-        <title>aiohttp page</title>
-    </head>
-    <body>
-        <div>
-            <h1><a href="/">My aiohttp server</a></h1>
-        </div>
-
-        <div>
-            <p>Date now: {{ current_date }}</p>
-            <p>Hello {{ username}}. </p>
-        </div>
-    </body>
-</html>
