--- code/05-templating/templates/index.html
@@ -1,22 +1,0 @@
-{% extends "base.html" %}
-{% block title %}
-Posts index
-{% endblock %}
-
-{% block content %}
-<h1>Posts</h1>
-<p>
-  <ul>
-    {% for post in posts %}
-    <li>
-      <a href="/{{ post.id }}">{{ post.title }}</a> {{ post.editor }}
-    </li>
-    {% endfor %}
-  </ul>
-</p>
-<hr>
-<p>
-  <a href="/new">New</a> 
-</p>
-
-{% endblock %}
