--- code/05-templating/templates/new.html
@@ -1,22 +1,0 @@
-{% extends "base.html" %}
-{% block title %}
-New Post
-{% endblock %}
-
-{% block content %}
-<h1>New Post</h1>
-<form action="/new" method="POST">
-  <p>
-    Title:
-    <input type="text" name="title" value="">
-  </p>
-  <p>Text:
-    <textarea type="text" name="text" value="" rows="10">
-  </p>
-  <input type="submit">
-</form>
-<hr>
-<p>
-  <a href="/">list</a> 
-</p>
-{% endblock %}
