--- code/05-templating/templates/view.html
@@ -1,15 +1,0 @@
-{% extends "base.html" %}
-{% block title %}
-Post {{ post.title }}
-{% endblock %}
-
-{% block content %}
-<h1>{{ post.title }}</h1>
-<p><pre> {{ post.text }}</pre> </p>
-<hr>
-<p><small> created by: {{ post.owner }}, last editor: {{ post.editor }} </small></p>
-<p>
-  <a href="/">list</a> 
-  <a href="/{{ post.id }}/edit">edit</a>
-</p>
-{% endblock %}
